Cast the 2nd parameter of ftruncate to off_t to work around bugs in some unix implementations.  Ticket #2425. (CVS 4089)
FossilOrigin-Name: 0b20a69609c64af922bedab381f7d075e9da4fc5

+++ src/os_unix.c
@@ -1271,7 +1271,7 @@ int sqlite3UnixSyncDirectory(const char *zDirname){
 static int unixTruncate(OsFile *id, i64 nByte){
   int rc;
   assert( id );
-  rc = ftruncate(((unixFile*)id)->h, nByte);
+  rc = ftruncate(((unixFile*)id)->h, (off_t)nByte);
   SimulateIOError( rc=1 );
   if( rc ){
     return SQLITE_IOERR_TRUNCATE;
